% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fun.name.R
\name{check_fun.name}
\alias{check_fun.name}
\title{Check usage function name.}
\usage{
check_fun.name(fun)
}
\arguments{
\item{fun}{Function, object.}
}
\value{
Prints a message: function bound with same name in other environments
or not.
}
\description{
Checks whether name of custom defined function is already bound
in other environments.
}
\examples{
check_fun.name(panel_hpfilter)

}
