% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_winsorize.R
\name{check_winsorize}
\alias{check_winsorize}
\title{Check outliers and subsequent winsorization.}
\usage{
check_winsorize(data, var, id, time, probs = c(0.05, 0.95),
  na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe.}

\item{var}{Variable of interest, object.}

\item{id}{Individual id for panel data, object.}

\item{time}{Time id for panel data, object.}

\item{probs}{Probability cut-off for winsorization. Default is 5\%.}

\item{na.rm}{Default is TRUE.}
}
\value{
List with three elements: graph with comparison pre/after
winsorization, dataframe with outliers and character vector with
individual ids outliers.
}
\description{
Check outliers and subsequent winsorization.
}
