% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_imf.R
\name{download_imf}
\alias{download_imf}
\title{Download IMF Data}
\usage{
download_imf(source, start, end, freq, seriename)
}
\arguments{
\item{source, start, end, freq, serie}{Character strings.}
}
\value{
Dataframe with country-id in iso2c, time-id and numeric value.
}
\description{
This function downloads a single serie for all countries available from
the IMF API and tidies it into dataframe format with country-id, time-id
and numeric value.
}
