% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_hpfilter.R
\name{panel_hpfilter}
\alias{panel_hpfilter}
\title{HP-filter for panel data.}
\usage{
panel_hpfilter(data, var, group, time, freq = 1600)
}
\arguments{
\item{data}{Dataframe.}

\item{var}{Variable of interest, object.}

\item{group}{Individuals ID panel, object.}

\item{time}{Time ID panel, object (character,numeric & date).}

\item{freq}{Frequency of HP filter, integer. Default is 1600.}
}
\value{
List with three elements. First element, original dataframe with
cyclical and trend component of variable, second element, list of graphs with
trend and variable and third element, list of ids not de-trended.
Warning message if some ids not de-trended.
}
\description{
Filters a single serie for different IDs in panel individually.
}
\examples{
panel_hpfilter(chinn_ito,ka_open,ccode,year, 1600)

}
