% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.bygroup.R
\name{plot_summary.bygroup}
\alias{plot_summary.bygroup}
\title{Plot cross sectional statistics in a panel.}
\usage{
plot_summary.bygroup(data, variable_to_summ, group, units,
  sum.stat = c("mean", "median", "sum", "n"), na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe.}

\item{variable_to_summ}{Variable of interest, numeric vector.}

\item{group}{Grouping variable, character or date vector.}

\item{sum.stat}{Summary statistic to plot, character. Default set to mean.}

\item{na.rm}{Default TRUE.}

\item{n}{Cross-sectional unit, character vector.}
}
\value{
List with two elements: plot and dataframe with values.
}
\description{
Plot average/median/cumulative value (and number of
cross-sectional units over which is performed the calculation) over time.
}
