% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{run_event.study}
\alias{run_event.study}
\title{Event study of variable of interest.}
\usage{
run_event.study(data, variable_of_interest, n_before, n_after,
  type = c("mean", "median"), level = 0.9)
}
\arguments{
\item{data}{Dataframe object.}

\item{variable_of_interest}{Character string. Name of indicator to track.}

\item{n_before}{Numeric (negative). Periods before the event.}

\item{n_after}{Numeric (positive). Periods after the event.}

\item{type}{Character. Summary static to track: default is "mean".}

\item{level}{Numeric. Default is 0.90.}
}
\value{
List with two elements - graph and dataframe on which latter based upon.
}
\description{
Performs an event study of a variable around the occurence of a certain
episode. Mean/median value before and after. CAUTION: build dummies before.
}
